--- Whistleblower/Views/Home/MailSent.cshtml
@@ -1,8 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "MailSent";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Ditt meddelande har skickats</h2>
-
