--- Whistleblower/Views/Home/ReportStatus.cshtml
@@ -1,8 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "ReportStatus";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Ärende ID</h2>
-
