+++ Whistleblower/Views/Login/LoginAdmin.cshtml
@@ -1,17 +1,19 @@
-﻿@model Whistleblower.Models.LoginLawyer
+﻿@model Whistleblower.Models.LoginAdmin
 
 @{
-    ViewBag.Title = "LoginLawyer";
+    ViewBag.Title = "LoginAdmin";
 }
-@using (Html.BeginForm("LoginLawyer", "Home", FormMethod.Post))
+
+@using (Html.BeginForm("LoginAdmin", "Login", FormMethod.Post))
 {
     @Html.AntiForgeryToken()
 
     <div class="form-horizontal">
         <div class="form-titel">
             <img src="~/images/omegapoint.png">
-            <h4>Login Lawyer</h4>
+            <h4>Login Admin</h4>
         </div>
+
         <hr />
         <div class="form-holder">
             <div class="form-body">
@@ -31,7 +33,7 @@
                         @Html.ValidationMessageFor(model => model.Password, "", new { @class = "text-danger" })
                         <input type="checkbox" onclick="myFunction()"> Visa lösenord
 
-                        <p id="text" style="display:none">VARNING! Caps lock är på.</p>
+                        <p id="text" class="text-danger" style="display:none">WARNING! Caps lock is ON.</p>
                         <script>
                             var input = document.getElementById("password-input");
                             var text = document.getElementById("text");
@@ -55,24 +57,19 @@
                         </script>
                     </div>
                 </div>
-
-            <div class="form-group">
-               <div class="col-md-offset-2 col-md-10">
-                    <input type="submit" value="Logga in" class="btn btn-default" />
-                    <div class="form-titel">
-                            @Html.ValidationMessage("LogOnError", new { @class = "text-danger" })
-                    </div>
+            <div class="col-md-offset-2 col-md-10">
+              <input type="submit" value="Logga in" class="btn btn-default" />
+               <div class="form-titel">
+                @Html.ValidationMessage("LogOnError", new { @class = "text-danger" })
                </div>
             </div>
+                
         </div>
     </div>
  </div>
 }
 
-<div>
-                @Html.ActionLink("Back to List", "Index")
-</div>
-
-            @section Scripts {
-                @Scripts.Render("~/bundles/jqueryval")
-            }
+@section Scripts 
+{
+   @Scripts.Render("~/bundles/jqueryval")
+}
