+++ Whistleblower/Views/Login/UserLogin.cshtml
@@ -6,7 +6,7 @@
 }
 
 
-@using (Html.BeginForm("UserLogin", "Home", FormMethod.Post ))
+@using (Html.BeginForm("UserLogin", "Login", FormMethod.Post ))
 {
     @Html.AntiForgeryToken()
