+++ Whistleblower/Views/Safebox/MailSent.cshtml
@@ -1,0 +1,8 @@
+﻿
+@{
+    ViewBag.Title = "MailSent";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<h2>Ditt meddelande har skickats</h2>
+
