+++ Whistleblower/Whistleblower.csproj
@@ -197,7 +197,9 @@
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
-    <Compile Include="Controllers\HomeController.cs" />
+    <Compile Include="Controllers\SafeboxController.cs" />
+    <Compile Include="Controllers\LoginController.cs" />
+    <Compile Include="Controllers\WhistleController.cs" />
     <Compile Include="Custom\DBHandler.cs" />
     <Compile Include="Custom\RequiredLabel.cs" />
     <Compile Include="Global.asax.cs">
@@ -246,18 +248,15 @@
     <Content Include="Views\_ViewStart.cshtml" />
     <Content Include="Views\Shared\Error.cshtml" />
     <Content Include="Views\Shared\_Layout.cshtml" />
-    <Content Include="Views\Home\About.cshtml" />
-    <Content Include="Views\Home\Contact.cshtml" />
-    <Content Include="Views\Home\Index.cshtml" />
-    <Content Include="Views\Home\Whistle.cshtml" />
-    <Content Include="Views\Home\Safebox.cshtml" />
-    <Content Include="Views\Home\SentMail.cshtml" />
-    <Content Include="Views\Home\ReportStatus.cshtml" />
-    <Content Include="Views\Home\LoginAdmin.cshtml" />
-    <Content Include="Views\Home\LoginLawyer.cshtml" />
-    <Content Include="Views\Home\MailSent.cshtml" />
-    <Content Include="Views\Home\UserLogin.cshtml" />
-    <Content Include="Views\Home\WhistleConfirm.cshtml" />
+    <Content Include="Views\Login\ReportStatus.cshtml" />
+    <Content Include="Views\Login\LoginAdmin.cshtml" />
+    <Content Include="Views\Login\LoginLawyer.cshtml" />
+    <Content Include="Views\Login\UserLogin.cshtml" />
+    <Content Include="Views\Whistle\Whistle.cshtml" />
+    <Content Include="Views\Whistle\WhistleConfirm.cshtml" />
+    <Content Include="Views\Safebox\MailSent.cshtml" />
+    <Content Include="Views\Safebox\Safebox.cshtml" />
+    <Content Include="Views\Safebox\SentMail.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
